+++ libs/components/src/lib/components/dropdowns/dropdown-host/dropdown-host.component.ts
@@ -100,6 +100,9 @@ export class PrizmDropdownHostComponent extends PrizmAbstractTestId implements A
   @Input() set isOpen(state: boolean) {
     this.isOpen$.next(state);
   }
+  get isOpen(): boolean {
+    return this.isOpen$.value;
+  }
 
   @Input() dropdownStyles: Record<string, string | number> = {};
 
@@ -167,12 +170,12 @@ export class PrizmDropdownHostComponent extends PrizmAbstractTestId implements A
 
   public close(): void {
     this.overlay?.close();
-    this.isOpenChange.emit(false);
+    if (this.isOpen) this.isOpenChange.emit(false);
   }
 
   public open(): void {
     this.overlay?.open();
-    this.isOpenChange.emit(true);
+    if (!this.isOpen) this.isOpenChange.emit(true);
   }
 
   private initOverlay(): void {